Fixed format (unsigned long long) in debug and logwrite calls

+++ check.c
@@ -248,22 +248,22 @@ void check_octet(struct checktype *pc, struct octet *octet, int level,
           exec_alarm(ip, octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), pc, 1);
           debug(1, "%s for %s is %lu - DoS\n", cp2str(pc->checkpoint),
                 printip(ip, 32, pc->by, pc->in),
-                octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
         }
         else
           logwrite("DoS %s %s: %lu %s", pc->in ? "to" : "from",
                    printip(ip, 32, pc->by, pc->in),
-                   octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), cp2str(pc->checkpoint));
+                   (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)), cp2str(pc->checkpoint));
         octet[i].alarmed = 1;
       } else if (octet[i].count >= (unsigned long long)pc->safelimit * (curtime - last_check)) {
         if (octet[i].alarmed)
           logwrite("DoS %s %s: %lu %s", pc->in ? "to" : "from",
                    printip(ip, 32, pc->by, pc->in),
-                   octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), cp2str(pc->checkpoint));
+                   (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)), cp2str(pc->checkpoint));
         else
           debug(1, "%s for %s is %lu - safe DoS\n", cp2str(pc->checkpoint),
                 printip(ip, 32, pc->by, pc->in),
-                octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
       } else {
         if (octet[i].alarmed) {
           exec_alarm(ip, octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), pc, 0);
@@ -272,7 +272,7 @@ void check_octet(struct checktype *pc, struct octet *octet, int level,
         if (octet[i].count)
           debug(2, "%s for %s is %lu - ok\n", cp2str(pc->checkpoint),
                 printip(ip, 32, pc->by, pc->in),
-                octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
       }
       octet[i].count = 0;
     } else if (octet[i].octet) {
@@ -287,13 +287,13 @@ void check_octet(struct checktype *pc, struct octet *octet, int level,
       if (octet[i].count >= (unsigned long long)pc->limit * (curtime - last_check)) {
         debug(1, "%s for %s is %lu - DoS, turning detailed stats on\n",
               cp2str(pc->checkpoint), printip(ip, 32, BYSRC, pc->in),
-              octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+              (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
         octet[i].used_time = curtime - expire_interval; /* remove on next check if no traffic */
         octet[i].octet = calloc(256, sizeof(struct octet));
       } else if (octet[i].count) {
         debug(2, "%s for %s is %lu - ok (no detailed stats)\n",
               cp2str(pc->checkpoint), printip(ip, 32, BYSRC, pc->in),
-              octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+              (unsigned long)(octet[i].count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
         octet[i].count = 0;
       }
     }
@@ -314,27 +314,27 @@ void check(void)
           exec_alarm((unsigned char *)&pc->ip, pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), pc, 1);
           debug(1, "%s for %s/%u is %lu - DoS\n", cp2str(pc->checkpoint),
                 inet_ntoa(*(struct in_addr *)&pc->ip), pc->preflen,
-                pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
         }
         else
           logwrite("DoS %s %s: %lu %s", pc->in ? "to" : "from",
                    printip((unsigned char *)&pc->ip, pc->preflen, pc->by, pc->in),
-                   pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), cp2str(pc->checkpoint));
+                   (unsigned long)(pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)), cp2str(pc->checkpoint));
         pc->alarmed = 1;
       } else if (pc->count >= (unsigned long long)pc->safelimit * (curtime - last_check)) {
         if (pc->alarmed)
           logwrite("DoS %s %s: %lu %s", pc->in ? "to" : "from",
                    printip((unsigned char *)&pc->ip, pc->preflen, pc->by, pc->in),
-                   pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), cp2str(pc->checkpoint));
+                   (unsigned long)(pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)), cp2str(pc->checkpoint));
         else
           debug(1, "%s for %s/%u is %lu - safe DoS\n", cp2str(pc->checkpoint),
                 inet_ntoa(*(struct in_addr *)&pc->ip), pc->preflen,
-                pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
       } else {
         if (pc->count)
           debug(2, "%s for %s/%u is %lu - ok\n", cp2str(pc->checkpoint),
                 inet_ntoa(*(struct in_addr *)&pc->ip), pc->preflen,
-                pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check));
+                (unsigned long)(pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check)));
 	if (pc->alarmed) {
           exec_alarm((unsigned char *)&pc->ip, pc->count * (pc->checkpoint == BPS ? 8 : 1) / (curtime - last_check), pc, 0);
           pc->alarmed = 0;
